+++ src/connectors/snowplow/actions/discover.js
@@ -65,6 +65,27 @@ export const discoverActions = {
       'description'
     ]
   },
+  'discover.unsave': {
+    eventType: 'engagement',
+    entityTypes: ['content', 'recommendation', 'slate', 'slateLineup', 'ui'],
+    eventData: {
+      uiType: 'button'
+    },
+    expects: [
+      'id',
+      'url',
+      'position',
+      'recommendationId',
+      'slateLineupId',
+      'slateLineupRequestId',
+      'slateLineupExperiment',
+      'slateId',
+      'slateRequestId',
+      'slateExperiment',
+      'displayName',
+      'description'
+    ]
+  },
   'discover.signup.impression': {
     eventType: 'impression',
     entityTypes: ['ui'],