+++ src/connectors/toasts/toast.state.js
@@ -5,6 +5,8 @@ import { MUTATION_SUCCESS } from 'actions'
 import { MUTATION_TAG_SUCCESS } from 'actions'
 import { MUTATION_DELETE_SUCCESS } from 'actions'
 
+import { ITEMS_UPSERT_SUCCESS } from 'actions'
+
 import { ITEMS_DELETE_SUCCESS } from 'actions'
 import { ITEMS_DELETE_FAILURE } from 'actions'
 
@@ -40,11 +42,6 @@ import { ITEMS_TAG_FAILURE } from 'actions'
 import { ADD_SHARE_SUCCESS } from 'actions'
 import { ADD_SHARE_FAILURE } from 'actions'
 
-import { PROFILE_ITEM_SAVE_SUCCESS } from 'actions'
-import { PROFILE_ITEM_SAVE_FAILURE } from 'actions'
-
-import { PROFILE_ITEM_DELETE_SUCCESS } from 'actions'
-import { PROFILE_ITEM_DELETE_FAILURE } from 'actions'
 
 /** ACTIONS
  --------------------------------------------------------------- */
@@ -95,15 +92,12 @@ export const actionToastsReducers = (state = initialState, action) => {
     case ITEMS_TAG_FAILURE:
     case ADD_SHARE_SUCCESS:
     case ADD_SHARE_FAILURE:
-    case PROFILE_ITEM_SAVE_SUCCESS:
-    case PROFILE_ITEM_SAVE_FAILURE:
-    case PROFILE_ITEM_DELETE_SUCCESS:
-    case PROFILE_ITEM_DELETE_FAILURE:
     case COLLECTIONS_SAVE_SUCCESS:
     case COLLECTION_PAGE_SAVE_SUCCESS:
     case DISCOVER_ITEMS_SAVE_SUCCESS:
     case ARTICLE_SAVE_SUCCESS:
     case MUTATION_SUCCESS:
+    case ITEMS_UPSERT_SUCCESS:
     case COPY_ITEM_URL: {
       const { actions, count, actionType = false } = action
       const itemCount = actions ? actions?.length : count