+++ src/containers/discover/topic/topic.state.js
@@ -1,27 +1,27 @@
-import { getDiscoverLineup } from 'common/api/queries/get-discover'
-import { DISCOVER_HYDRATE } from 'actions'
+import { getTopicLineup } from 'common/api/queries/get-topic-lineup'
+import { TOPIC_HYDRATE } from 'actions'
 import { HYDRATE } from 'actions'
 
 /** ACTIONS
  --------------------------------------------------------------- */
-export const hydrateDiscover = (itemIds) => ({ type: DISCOVER_HYDRATE, itemIds }) //prettier-ignore
+export const hydrateTopic = ({ topic, itemIds }) => ({ type: TOPIC_HYDRATE, topic, itemIds })
 
 /** REDUCERS
  --------------------------------------------------------------- */
-const initialState = []
+const initialState = {}
 
-export const pageDiscoverIdsReducers = (state = initialState, action) => {
+export const pageTopicReducers = (state = initialState, action) => {
   switch (action.type) {
-    case DISCOVER_HYDRATE: {
-      const { itemIds } = action
-      return itemIds
+    case TOPIC_HYDRATE: {
+      const { topic, itemIds } = action
+      return { ...state, [topic]: itemIds }
     }
 
     // SPECIAL HYDRATE:  This is sent from the next-redux wrapper and
     // it represents the state used to build the page on the server.
     case HYDRATE: {
-      const { pageDiscoverIds } = action.payload
-      return pageDiscoverIds
+      const { pageTopic } = action.payload
+      return pageTopic
     }
 
     default:
@@ -33,17 +33,17 @@ export const pageDiscoverIdsReducers = (state = initialState, action) => {
  --------------------------------------------------------------- */
 
 /**
- * fetchDiscoverData
+ * fetchTopicData
  * Make and async request for a Pocket v3 feed and return best data
  * @return items {array} An array of derived items
  */
-export async function fetchDiscoverData({ locale }) {
+export async function fetchTopicData(topic) {
   try {
-    const { itemsById, itemIds } = await getDiscoverLineup({ locale })
-
-    return { itemIds, itemsById }
+    const { itemsById, itemIds } = await getTopicLineup(topic, 30)
+    return { topic, itemsById, itemIds }
   } catch (error) {
     //TODO: adjust this once error reporting strategy is defined.
-    console.warn('discover.state', error)
+    console.warn('topic-pages.topic-data.state', error)
+    return false
   }
 }